requirements.txt
tensorflow
dvc
pandas
Flask
numpy
ipykernel
seaborn
matplotlib
pyYAML
python-box==6.0.2
ensure==1.0.2
joblib
nltk
spacy
emoji
gensim
scikit-learn
gunicorn
-e .
